requirements.txt
asgiref==3.2.7
boto3==1.12.41
botocore==1.15.41
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.0.4
Django==3.0.5
django-allauth==0.41.0
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
drf-extensions==0.6.0
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.1.0
